requirements.txt
Django==1.10.6
django-cors-headers==2.0.2
djangorestframework==3.6.2
wheel==0.24.0
